Add default settings
With this change, one can control the default user settings in the
application without having to modify slicer's code. It works by checking
the content of the user setting file. If the setting variables in the
default settings file are not in the user settings file(for example
the first time the application is started) then the
setting variables in DefaultSettings.ini are being copied into the
user settings file.